+++ src/strtoIdd.c
@@ -38,7 +38,7 @@ int strtoIdd(const char *s, char **sp, double *f0, double *f1)
 #else
 	static FPI fpi = {106, 1 - 1023 - 53 + 1, 2046 - 1023 - 106 + 1, 1, 0};
 #endif
-	Long exp[2];
+	int32_t exp[2];
 	Bigint* B[2];
 	int k;
 	int rv[2];
@@ -46,20 +46,20 @@ int strtoIdd(const char *s, char **sp, double *f0, double *f1)
 	B[0] = Balloc(2);
 	B[0]->wds = 4;
 	k = strtoIg(s, sp, &fpi, exp, B, rv);
-	ULtodd((ULong*)f0, B[0]->x, exp[0], rv[0]);
+	ULtodd((uint32_t*)f0, B[0]->x, exp[0], rv[0]);
 	Bfree(B[0]);
 
 	if(B[1])
 	{
-		ULtodd((ULong*)f1, B[1]->x, exp[1], rv[1]);
+		ULtodd((uint32_t*)f1, B[1]->x, exp[1], rv[1]);
 		Bfree(B[1]);
 	}
 	else
 	{
-		((ULong*)f1)[0] = ((ULong*)f0)[0];
-		((ULong*)f1)[1] = ((ULong*)f0)[1];
-		((ULong*)f1)[2] = ((ULong*)f0)[2];
-		((ULong*)f1)[3] = ((ULong*)f0)[3];
+		((uint32_t*)f1)[0] = ((uint32_t*)f0)[0];
+		((uint32_t*)f1)[1] = ((uint32_t*)f0)[1];
+		((uint32_t*)f1)[2] = ((uint32_t*)f0)[2];
+		((uint32_t*)f1)[3] = ((uint32_t*)f0)[3];
 	}
 
 	return k;
