+++ src/strtodnrp.c
@@ -40,12 +40,12 @@ THIS SOFTWARE.
 double strtod(const char* s, char** sp)
 {
 	static FPI fpi = {53, 1 - 1023 - 53 + 1, 2046 - 1023 - 53 + 1, 1, SI};
-	ULong bits[2];
-	Long exp;
+	uint32_t bits[2];
+	int32_t exp;
 	int k;
 	union
 	{
-		ULong L[2];
+		uint32_t L[2];
 		double d;
 	} u;
