+++ src/strtof.c
@@ -34,12 +34,12 @@ THIS SOFTWARE.
 float strtof(const char *s, char **sp)
 {
 	static FPI fpi = {24, 1 - 127 - 24 + 1, 254 - 127 - 24 + 1, 1, SI};
-	ULong bits[1];
-	Long exp;
+	uint32_t bits[1];
+	int32_t exp;
 	int k;
 	union
 	{
-		ULong L[1];
+		uint32_t L[1];
 		float f;
 	} u;
